requirements.txt
pysrt
webvtt-py 
moviepy
imagemagick